Change test path to be independent of checkout name
Working relative to the (changing) current working directory in Java sucks in
general.

It is really hard to do without intentionally spawning another VM with changed
environment. Setting System property 'user.dir' sucks because it is not a
reliable CWD change; some stdlib Elements respect it; others don’t.

This does not go down from the CWD, but at least it still checks .. as a path
element FWIW.