Convert io package
This is the first time a hard problem with automatic translation occurred -
jinxed it with the last commit!

The translator forgot the @JvmStatic annotation on Directories::copyDir, which
meant that groovy code expecting Java call semantics could not find it.

I guess this is caused by groovy not being statically analyzed for usages
because of its dynamically typed nature.

Curious because navigation to Java/Kotlin callees from Groovy works in the
editor.