Convert util package
What a bloody nightmare. The conversion ended with a bang.

ChecksumCalculator:

Kotlin support for bit fiddling is surprisingly poor. Ended up at a probably not
too inefficient implementation. Not enjoyable, and finally this would pull the
...Experimental annotations all through the project including users of the class.

I just left it in Java, but left my implementation in ChecksumCalculatorKt.kt.

For that, I had to duplicate the Tests as well and wrap both Kotlin Methods in
one since using the unsigned byte (UByte [& ...Array]) type only works within
Kotlin calls; when you return it outside (in this case, Groovy), it gets treated
as regular byte[] and then cannot be passed back into Kotlin as a UByteArray.

This was one of only two times during the whole conversion that anything about
Kotlin type system semantics was confusing to me, and the other one was a
harmless bit of convenience magic. This, however, was a doozy! Total mess.

StringUtils:

I *really* jinxed it with talking about the new flawless Java -> Kotlin
conversion. Once again the @JvmStatic annotations were missing on the methods in
this class, breaking the tests. This seems to once again be caused by the
converter not detecting dependencies in the (dynamic) Groovy context. Earlier in
the conversion process this did not come up because there were other (Java)
users so @JvmStatic was added; now that all other users are Kotlin and the
methods are not used in any "regular" static context any more, this process
bombs. One to watch out for!

Note: The Java impl of trimmedJoin does not seem to do what it says on the can -
it doesn’t trim delimiter from the ends but rather every char *in* delimiter
from anywhere in the string. The Kotlin impl. changes the semantics to take the
whole delimiter from both ends of the string only. It passes all previously
present tests.